ark_r1cs_std::bits::boolean

Struct AllocatedBool

pub struct AllocatedBool<F: Field> { /* private fields */ }

Represents a variable in the constraint system which is guaranteed to be either zero or one.

In general, one should prefer using Boolean instead of AllocatedBool, as Boolean offers better support for constant values, and implements more traits.

Implementations

impl<F: Field> AllocatedBool<F>

pub fn value(&self) -> Result<bool, SynthesisError>

Get the assigned value for self.

pub fn variable(&self) -> Variable

Get the R1CS variable for self.

pub fn xor(&self, b: &Self) -> Result<Self, SynthesisError>

Performs an XOR operation over the two operands, returning an AllocatedBool.

pub fn and(&self, b: &Self) -> Result<Self, SynthesisError>

Performs an AND operation over the two operands, returning an AllocatedBool.

pub fn or(&self, b: &Self) -> Result<Self, SynthesisError>

Performs an OR operation over the two operands, returning an AllocatedBool.

pub fn and_not(&self, b: &Self) -> Result<Self, SynthesisError>

Calculates a AND (NOT b).

pub fn nor(&self, b: &Self) -> Result<Self, SynthesisError>

Calculates (NOT a) AND (NOT b).