ark_r1cs_std

Module fields

This module implements gadgets related to field arithmetic.

Modules

cubic_extension

This module contains a generic implementation of cubic extension field variables. That is, it implements the R1CS equivalent of ark_ff::CubicExtField.

fp

This module contains a generic implementation of prime field variables. That is, it implements the R1CS equivalent of ark_ff::Fp*.

fp2

This module contains a generic implementation of the degree-2 tower extension field. That is, it implements the R1CS equivalent of ark_ff::Fp2

fp3

This module contains a generic implementation of the degree-3 tower extension field. That is, it implements the R1CS equivalent of ark_ff::Fp3

fp4

This module contains a generic implementation of the degree-4 tower extension field. That is, it implements the R1CS equivalent of ark_ff::Fp4

fp6_2over3

This module contains a generic implementation of the degree-6 tower extension field. That is, it implements the R1CS equivalent of ark_ff::fp6_2over3::Fp6

fp6_3over2

This module contains a generic implementation of the degree-6 tower extension field. That is, it implements the R1CS equivalent of ark_ff::fp6_3over2::Fp6

fp12

This module contains a generic implementation of the degree-12 tower extension field. That is, it implements the R1CS equivalent of ark_ff::Fp12

nonnative

This module contains a generic implementation of “nonnative” prime field variables. It emulates Fp arithmetic using Fq operations, where p != q.

quadratic_extension

This module contains a generic implementation of quadratic extension field variables. That is, it implements the R1CS equivalent of ark_ff::QuadExtField.

Traits

FieldOpsBounds

This trait is a hack used to work around the lack of implied bounds.

FieldVar

A variable representing a field. Corresponds to the native type F.