ark_r1cs_std::fields::cubic_extension

Struct CubicExtVar

pub struct CubicExtVar<BF: FieldVar<P::BaseField, P::BasePrimeField>, P: CubicExtVarConfig<BF>>
where
    for<'a> &'a BF: FieldOpsBounds<'a, P::BaseField, BF>,
{
    pub c0: BF,
    pub c1: BF,
    pub c2: BF,
    /* private fields */
}

This struct is the R1CS equivalent of the cubic extension field type in ark-ff, i.e. ark_ff::CubicExtField.

Fields
c0: BF

The zero-th coefficient of this field element.

c1: BF

The first coefficient of this field element.

c2: BF

The second coefficient of this field element.

Implementations

impl<BF: FieldVar<P::BaseField, P::BasePrimeField>, P: CubicExtVarConfig<BF>> CubicExtVar<BF, P>
where for<'a> &'a BF: FieldOpsBounds<'a, P::BaseField, BF>,

pub fn new(c0: BF, c1: BF, c2: BF) -> Self

Constructs a CubicExtVar from the underlying coefficients.

pub fn mul_base_field_by_nonresidue(fe: &BF) -> Result<BF, SynthesisError>

Multiplies a variable of the base field by the cubic nonresidue P::NONRESIDUE that is used to construct the extension field.

pub fn mul_by_base_field_constant(&self, fe: P::BaseField) -> Self

Multiplies self by a constant from the base field.

pub fn mul_assign_by_base_field_constant(&mut self, fe: P::BaseField)

Sets self = self.mul_by_base_field_constant(fe).

impl<P: Fp6Config> CubicExtVar<QuadExtVar<FpVar<<<P as Fp6Config>::Fp2Config as Fp2Config>::Fp>, Fp2ConfigWrapper<<P as Fp6Config>::Fp2Config>>, Fp6ConfigWrapper<P>>

pub fn mul_by_0_c1_0( &self, c1: &Fp2Var<P::Fp2Config>, ) -> Result<Self, SynthesisError>

Multiplies self by a sparse element which has c0 == c2 == zero.

pub fn mul_by_c0_c1_0( &self, c0: &Fp2Var<P::Fp2Config>, c1: &Fp2Var<P::Fp2Config>, ) -> Result<Self, SynthesisError>

Multiplies self by a sparse element which has c2 == zero.