ark_r1cs_std::fields::fp4
Type Alias Fp4Var

pub type Fp4Var<P> = QuadExtVar<Fp2Var<<P as Fp4Config>::Fp2Config>, Fp4ConfigWrapper<P>>;

A quartic extension field constructed as the tower of a quadratic extension over a quadratic extension field. This is the R1CS equivalent of ark_ff::Fp4<P>.

Aliased Type

struct Fp4Var<P> {
    pub c0: QuadExtVar<FpVar<<<P as Fp4Config>::Fp2Config as Fp2Config>::Fp>, Fp2ConfigWrapper<<P as Fp4Config>::Fp2Config>>,
    pub c1: QuadExtVar<FpVar<<<P as Fp4Config>::Fp2Config as Fp2Config>::Fp>, Fp2ConfigWrapper<<P as Fp4Config>::Fp2Config>>,
    /* private fields */
}

Fields
c0: QuadExtVar<FpVar<<<P as Fp4Config>::Fp2Config as Fp2Config>::Fp>, Fp2ConfigWrapper<<P as Fp4Config>::Fp2Config>>

The zero-th coefficient of this field element.

c1: QuadExtVar<FpVar<<<P as Fp4Config>::Fp2Config as Fp2Config>::Fp>, Fp2ConfigWrapper<<P as Fp4Config>::Fp2Config>>

The first coefficient of this field element.