ark_r1cs_std::fields::fp6_2over3
Type Alias Fp6Var

pub type Fp6Var<P> = QuadExtVar<Fp3Var<<P as Fp6Config>::Fp3Config>, Fp6ConfigWrapper<P>>;

A sextic extension field constructed as the tower of a quadratic extension over a cubic extension field. This is the R1CS equivalent of ark_ff::fp6_2over3::Fp6<P>.

Aliased Type

struct Fp6Var<P> {
    pub c0: CubicExtVar<FpVar<<<P as Fp6Config>::Fp3Config as Fp3Config>::Fp>, Fp3ConfigWrapper<<P as Fp6Config>::Fp3Config>>,
    pub c1: CubicExtVar<FpVar<<<P as Fp6Config>::Fp3Config as Fp3Config>::Fp>, Fp3ConfigWrapper<<P as Fp6Config>::Fp3Config>>,
    /* private fields */
}

Fields
c0: CubicExtVar<FpVar<<<P as Fp6Config>::Fp3Config as Fp3Config>::Fp>, Fp3ConfigWrapper<<P as Fp6Config>::Fp3Config>>

The zero-th coefficient of this field element.

c1: CubicExtVar<FpVar<<<P as Fp6Config>::Fp3Config as Fp3Config>::Fp>, Fp3ConfigWrapper<<P as Fp6Config>::Fp3Config>>

The first coefficient of this field element.