ark_r1cs_std::fields::fp6_3over2
Type Alias Fp6Var

pub type Fp6Var<P> = CubicExtVar<Fp2Var<<P as Fp6Config>::Fp2Config>, Fp6ConfigWrapper<P>>;

A sextic extension field constructed as the tower of a cubic extension over a quadratic extension field. This is the R1CS equivalent of ark_ff::fp6_3over3::Fp6<P>.

Aliased Type

struct Fp6Var<P> {
    pub c0: QuadExtVar<FpVar<<<P as Fp6Config>::Fp2Config as Fp2Config>::Fp>, Fp2ConfigWrapper<<P as Fp6Config>::Fp2Config>>,
    pub c1: QuadExtVar<FpVar<<<P as Fp6Config>::Fp2Config as Fp2Config>::Fp>, Fp2ConfigWrapper<<P as Fp6Config>::Fp2Config>>,
    pub c2: QuadExtVar<FpVar<<<P as Fp6Config>::Fp2Config as Fp2Config>::Fp>, Fp2ConfigWrapper<<P as Fp6Config>::Fp2Config>>,
    /* private fields */
}

Fields
c0: QuadExtVar<FpVar<<<P as Fp6Config>::Fp2Config as Fp2Config>::Fp>, Fp2ConfigWrapper<<P as Fp6Config>::Fp2Config>>

The zero-th coefficient of this field element.

c1: QuadExtVar<FpVar<<<P as Fp6Config>::Fp2Config as Fp2Config>::Fp>, Fp2ConfigWrapper<<P as Fp6Config>::Fp2Config>>

The first coefficient of this field element.

c2: QuadExtVar<FpVar<<<P as Fp6Config>::Fp2Config as Fp2Config>::Fp>, Fp2ConfigWrapper<<P as Fp6Config>::Fp2Config>>

The second coefficient of this field element.

Implementations

impl<P: Fp6Config> Fp6Var<P>

pub fn mul_by_0_c1_0( &self, c1: &Fp2Var<P::Fp2Config>, ) -> Result<Self, SynthesisError>

Multiplies self by a sparse element which has c0 == c2 == zero.

pub fn mul_by_c0_c1_0( &self, c0: &Fp2Var<P::Fp2Config>, c1: &Fp2Var<P::Fp2Config>, ) -> Result<Self, SynthesisError>

Multiplies self by a sparse element which has c2 == zero.