ark_r1cs_std::fields::nonnative

Enum NonNativeFieldMulResultVar

pub enum NonNativeFieldMulResultVar<TargetField: PrimeField, BaseField: PrimeField> {
    Constant(TargetField),
    Var(AllocatedNonNativeFieldMulResultVar<TargetField, BaseField>),
}

An intermediate representation especially for the result of a multiplication, containing more limbs. It is intended for advanced usage to improve the efficiency.

That is, instead of calling mul, one can call mul_without_reduce to obtain this intermediate representation, which can still be added. Then, one can call reduce to reduce it back to NonNativeFieldVar. This may help cut the number of reduce operations.

Variants

Constant(TargetField)

as a constant

Var(AllocatedNonNativeFieldMulResultVar<TargetField, BaseField>)

as an allocated gadget

Implementations

impl<TargetField: PrimeField, BaseField: PrimeField> NonNativeFieldMulResultVar<TargetField, BaseField>

pub fn zero() -> Self

Create a zero NonNativeFieldMulResultVar (used for additions)

pub fn constant(v: TargetField) -> Self

Create an NonNativeFieldMulResultVar from a constant

pub fn reduce(&self) -> R1CSResult<NonNativeFieldVar<TargetField, BaseField>>

Reduce the NonNativeFieldMulResultVar back to NonNativeFieldVar