ark_r1cs_std::fields::nonnative
Module params

Utilities for sampling parameters for non-native field gadgets

BaseField: the constraint field

TargetField: the field being simulated

num_limbs: how many limbs are used

bits_per_limb: the size of the limbs

Enums

OptimizationType

The type of optimization target for the parameters searching

Functions

find_parameters

A function to search for parameters for nonnative field gadgets

get_params

Obtain the parameters from a ConstraintSystem’s cache or generate a new one