ark_r1cs_std::fields::nonnative

Struct AllocatedNonNativeFieldVar

pub struct AllocatedNonNativeFieldVar<TargetField: PrimeField, BaseField: PrimeField> {
    pub cs: ConstraintSystemRef<BaseField>,
    pub limbs: Vec<FpVar<BaseField>>,
    pub num_of_additions_over_normal_form: BaseField,
    pub is_in_the_normal_form: bool,
    /* private fields */
}

The allocated version of NonNativeFieldVar (introduced below)

Fields
cs: ConstraintSystemRef<BaseField>

Constraint system reference

limbs: Vec<FpVar<BaseField>>

The limbs, each of which is a BaseField gadget.

num_of_additions_over_normal_form: BaseField

Number of additions done over this gadget, using which the gadget decides when to reduce.

is_in_the_normal_form: bool

Whether the limb representation is the normal form (using only the bits specified in the parameters, and the representation is strictly within the range of TargetField).

Implementations

impl<TargetField: PrimeField, BaseField: PrimeField> AllocatedNonNativeFieldVar<TargetField, BaseField>

pub fn cs(&self) -> ConstraintSystemRef<BaseField>

Return cs

pub fn limbs_to_value( limbs: Vec<BaseField>, optimization_type: OptimizationType, ) -> TargetField

Obtain the value of limbs

pub fn value(&self) -> R1CSResult<TargetField>

Obtain the value of a nonnative field element

pub fn constant( cs: ConstraintSystemRef<BaseField>, value: TargetField, ) -> R1CSResult<Self>

Obtain the nonnative field element of a constant value

pub fn one(cs: ConstraintSystemRef<BaseField>) -> R1CSResult<Self>

Obtain the nonnative field element of one

pub fn zero(cs: ConstraintSystemRef<BaseField>) -> R1CSResult<Self>

Obtain the nonnative field element of zero

pub fn add(&self, other: &Self) -> R1CSResult<Self>

Add a nonnative field element

pub fn add_constant(&self, other: &TargetField) -> R1CSResult<Self>

Add a constant

pub fn sub_without_reduce(&self, other: &Self) -> R1CSResult<Self>

Subtract a nonnative field element, without the final reduction step

pub fn sub(&self, other: &Self) -> R1CSResult<Self>

Subtract a nonnative field element

pub fn sub_constant(&self, other: &TargetField) -> R1CSResult<Self>

Subtract a constant

pub fn mul(&self, other: &Self) -> R1CSResult<Self>

Multiply a nonnative field element

pub fn mul_constant(&self, other: &TargetField) -> R1CSResult<Self>

Multiply a constant

pub fn negate(&self) -> R1CSResult<Self>

Compute the negate of a nonnative field element

pub fn inverse(&self) -> R1CSResult<Self>

Compute the inverse of a nonnative field element

pub fn get_limbs_representations( elem: &TargetField, optimization_type: OptimizationType, ) -> R1CSResult<Vec<BaseField>>

Convert a TargetField element into limbs (not constraints) This is an internal function that would be reused by a number of other functions

pub fn get_limbs_representations_from_big_integer( elem: &<TargetField as PrimeField>::BigInt, optimization_type: OptimizationType, ) -> R1CSResult<Vec<BaseField>>

Obtain the limbs directly from a big int

pub fn mul_without_reduce( &self, other: &Self, ) -> R1CSResult<AllocatedNonNativeFieldMulResultVar<TargetField, BaseField>>

for advanced use, multiply and output the intermediate representations (without reduction) This intermediate representations can be added with each other, and they can later be reduced back to the NonNativeFieldVar.

pub fn new_witness_with_le_bits<T: Borrow<TargetField>>( cs: impl Into<Namespace<BaseField>>, f: impl FnOnce() -> Result<T, SynthesisError>, ) -> R1CSResult<(Self, Vec<Boolean<BaseField>>)>

Allocates a new non-native field witness with value given by the function f. Enforces that the field element has value in [0, modulus), and returns the bits of its binary representation. The bits are in little-endian (i.e., the bit at index 0 is the LSB) and the bit-vector is empty in non-witness allocation modes.