ark_r1cs_std::fields::quadratic_extension

Struct QuadExtVar

pub struct QuadExtVar<BF: FieldVar<P::BaseField, P::BasePrimeField>, P: QuadExtVarConfig<BF>>
where
    for<'a> &'a BF: FieldOpsBounds<'a, P::BaseField, BF>,
{
    pub c0: BF,
    pub c1: BF,
    /* private fields */
}

This struct is the R1CS equivalent of the quadratic extension field type in ark-ff, i.e. ark_ff::QuadExtField.

Fields
c0: BF

The zero-th coefficient of this field element.

c1: BF

The first coefficient of this field element.

Implementations

impl<BF: FieldVar<P::BaseField, P::BasePrimeField>, P: QuadExtVarConfig<BF>> QuadExtVar<BF, P>
where for<'a> &'a BF: FieldOpsBounds<'a, P::BaseField, BF>,

pub fn new(c0: BF, c1: BF) -> Self

Constructs a QuadExtVar from the underlying coefficients.

pub fn mul_base_field_by_nonresidue(fe: &BF) -> Result<BF, SynthesisError>

Multiplies a variable of the base field by the quadratic nonresidue P::NONRESIDUE that is used to construct the extension field.

pub fn mul_by_base_field_constant(&self, fe: P::BaseField) -> Self

Multiplies self by a constant from the base field.

pub fn mul_assign_by_base_field_constant(&mut self, fe: P::BaseField)

Sets self = self.mul_by_base_field_constant(fe).

pub fn unitary_inverse(&self) -> Result<Self, SynthesisError>

This is only to be used when the element is known to be in the cyclotomic subgroup.

pub fn cyclotomic_exp( &self, exponent: impl AsRef<[u64]>, ) -> Result<Self, SynthesisError>
where Self: FieldVar<QuadExtField<P>, P::BasePrimeField>,

This is only to be used when the element is known to be in the cyclotomic subgroup.

impl<P: Fp12Config> QuadExtVar<CubicExtVar<QuadExtVar<FpVar<<<<P as Fp12Config>::Fp6Config as Fp6Config>::Fp2Config as Fp2Config>::Fp>, Fp2ConfigWrapper<<<P as Fp12Config>::Fp6Config as Fp6Config>::Fp2Config>>, Fp6ConfigWrapper<<P as Fp12Config>::Fp6Config>>, Fp12ConfigWrapper<P>>

pub fn mul_by_014( &self, c0: &Fp2Var<<<P as Fp12Config>::Fp6Config as Fp6Config>::Fp2Config>, c1: &Fp2Var<<<P as Fp12Config>::Fp6Config as Fp6Config>::Fp2Config>, d1: &Fp2Var<<<P as Fp12Config>::Fp6Config as Fp6Config>::Fp2Config>, ) -> Result<Self, SynthesisError>

Multiplies by a sparse element of the form (c0 = (c0, c1, 0), c1 = (0, d1, 0)).

pub fn mul_by_034( &self, c0: &Fp2Var<<<P as Fp12Config>::Fp6Config as Fp6Config>::Fp2Config>, d0: &Fp2Var<<<P as Fp12Config>::Fp6Config as Fp6Config>::Fp2Config>, d1: &Fp2Var<<<P as Fp12Config>::Fp6Config as Fp6Config>::Fp2Config>, ) -> Result<Self, SynthesisError>

Multiplies by a sparse element of the form (c0 = (c0, 0, 0), c1 = (d0, d1, 0)).

pub fn cyclotomic_square(&self) -> Result<Self, SynthesisError>

Squares self when self is in the cyclotomic subgroup.

pub fn optimized_cyclotomic_exp( &self, exponent: impl AsRef<[u64]>, ) -> Result<Self, SynthesisError>

Like Self::cyclotomic_exp, but additionally uses cyclotomic squaring.