ark_r1cs_std::groups::curves
Module short_weierstrass

This module generically implements arithmetic for Short Weierstrass elliptic curves by following the complete formulae of [Renes, Costello, Batina 2015].

Modules

bls12

This module provides a generic implementation of G1 and G2 for the [[BLS12]](https://eprint.iacr.org/2002/088.pdf) family of bilinear groups.

mnt4

This module provides a generic implementation of G1 and G2 for the [[MNT4]](https://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.20.8113&rep=rep1&type=pdf) family of bilinear groups.

mnt6

This module provides a generic implementation of G1 and G2 for the [[MNT6]](https://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.20.8113&rep=rep1&type=pdf) family of bilinear groups.

non_zero_affine

This module provides a generic implementation of elliptic curve operations for points on short-weierstrass curves in affine coordinates that are not equal to zero.

Structs

AffineVar

An affine representation of a curve point.

ProjectiveVar

An implementation of arithmetic for Short Weierstrass curves that relies on the complete formulae derived in the paper of [Renes, Costello, Batina 2015].