ark_r1cs_std::groups::curves::short_weierstrass
Module bls12

This module provides a generic implementation of G1 and G2 for the [[BLS12]](https://eprint.iacr.org/2002/088.pdf) family of bilinear groups.

Structs

G1PreparedVar

Represents the cached precomputation that can be performed on a G1 element which enables speeding up pairing computation.

G2PreparedVar

Represents the cached precomputation that can be performed on a G2 element which enables speeding up pairing computation.

Type Aliases

G1AffineVar

Represents an affine point on G1. Should be used only for comparison and when a canonical representation of a point is required, and not for arithmetic.

G1Var

Represents a projective point in G1.

G2AffineVar

Represents an affine point on G2. Should be used only for comparison and when a canonical representation of a point is required, and not for arithmetic.

G2Var

Represents a projective point in G2.