ark_r1cs_std::groups::curves::short_weierstrass
Module non_zero_affine

This module provides a generic implementation of elliptic curve operations for points on short-weierstrass curves in affine coordinates that are not equal to zero.

Note: this module is unsafe in general: it can synthesize unsatisfiable or underconstrained constraint systems when a represented point is equal to zero. The ProjectiveVar gadget is the recommended way of working with elliptic curve points.

Structs

NonZeroAffineVar

An affine representation of a prime order curve point that is guaranteed to not be the point at infinity.