ark_r1cs_std::groups::curves::twisted_edwards

Struct MontgomeryAffineVar

pub struct MontgomeryAffineVar<P: TEModelParameters, F: FieldVar<P::BaseField, <P::BaseField as Field>::BasePrimeField>>
where
    for<'a> &'a F: FieldOpsBounds<'a, P::BaseField, F>,
{
    pub x: F,
    pub y: F,
    /* private fields */
}

An implementation of arithmetic for Montgomery curves that relies on incomplete addition formulae for the affine model, as outlined in the EFD.

This is intended for use primarily for implementing efficient multi-scalar-multiplication in the Bowe-Hopwood-Pedersen hash.

Fields
x: F

The x-coordinate.

y: F

The y-coordinate.

Implementations

impl<P: TEModelParameters, F: FieldVar<P::BaseField, <P::BaseField as Field>::BasePrimeField>> MontgomeryAffineVar<P, F>
where for<'a> &'a F: FieldOpsBounds<'a, P::BaseField, F>,

pub fn new(x: F, y: F) -> Self

Constructs Self from an (x, y) coordinate pair.

pub fn from_edwards_to_coords( p: &TEAffine<P>, ) -> Result<(P::BaseField, P::BaseField), SynthesisError>

Converts a Twisted Edwards curve point to coordinates for the corresponding affine Montgomery curve point.

pub fn new_witness_from_edwards( cs: ConstraintSystemRef<<P::BaseField as Field>::BasePrimeField>, p: &TEAffine<P>, ) -> Result<Self, SynthesisError>

Converts a Twisted Edwards curve point to coordinates for the corresponding affine Montgomery curve point.

pub fn into_edwards(&self) -> Result<AffineVar<P, F>, SynthesisError>

Converts self into a Twisted Edwards curve point variable.