ark_r1cs_std::poly::domain

Struct Radix2DomainVar

pub struct Radix2DomainVar<F: PrimeField> {
    pub gen: F,
    pub dim: u64,
    /* private fields */
}

Defines an evaluation domain over a prime field. The domain is a coset of size 1<<dim.

Native code corresponds to ark-poly::univariate::domain::radix2, but ark-poly only supports subgroup for now.

Fields
gen: F

generator of subgroup g

dim: u64

dimension of evaluation domain, which is log2(size of coset)

Implementations

impl<F: PrimeField> Radix2DomainVar<F>

pub fn new( gen: F, dimension: u64, offset: FpVar<F>, ) -> Result<Self, SynthesisError>

Construct an evaluation domain with the given offset.

pub fn offset(&self) -> &FpVar<F>

What is the offset of self?

impl<F: PrimeField> Radix2DomainVar<F>

pub fn order(&self) -> usize

order of the domain

pub fn elements(&self) -> Vec<FpVar<F>>

Returns offset, offsetg, offsetg^2, โฆ, offset*g^{coset_size}

pub fn size(&self) -> u64

Size of the domain

pub fn query_position_to_coset_elements( &self, query_pos: &[Boolean<F>], coset_dim: u64, ) -> Result<Vec<FpVar<F>>, SynthesisError>

For domain h<g> with dimension n, position represented by query_pos in big endian form, returns all points of h*g^{position}<g^{2^{n-coset_dim}}>. The result is the query coset at index query_pos for the FRI protocol.

Panics

This function panics when query_pos.len() != coset_dim or query_pos.len() != self.dim.

pub fn query_position_to_coset( &self, query_pos: &[Boolean<F>], coset_dim: u64, ) -> Result<Self, SynthesisError>

For domain h<g> with dimension n, position represented by query_pos in big endian form, returns all points of h*g^{position}<g^{n-query_pos.len()}>

Panics

This function panics when query_pos.len() < log2_num_cosets.