ark_r1cs_std::poly::evaluations::univariate

Struct EvaluationsVar

pub struct EvaluationsVar<F: PrimeField> {
    pub evals: Vec<FpVar<F>>,
    pub lagrange_interpolator: Option<LagrangeInterpolator<F>>,
    /* private fields */
}

Stores a UV polynomial in evaluation form.

Fields
evals: Vec<FpVar<F>>

Evaluations of univariate polynomial over domain

lagrange_interpolator: Option<LagrangeInterpolator<F>>

Optional Lagrange Interpolator. Useful for lagrange interpolation.

Implementations

impl<F: PrimeField> EvaluationsVar<F>

pub fn from_vec_and_domain( evaluations: Vec<FpVar<F>>, domain: Radix2DomainVar<F>, interpolate: bool, ) -> Self

Construct Self from evaluations and a domain. interpolate indicates if user wants to interpolate this polynomial using lagrange interpolation.

pub fn generate_interpolation_cache(&mut self)

Precompute necessary calculation for lagrange interpolation and mark it ready to interpolate

pub fn interpolate_and_evaluate( &self, interpolation_point: &FpVar<F>, ) -> Result<FpVar<F>, SynthesisError>

Returns constraints for Interpolating and evaluating at interpolation_point