ark_r1cs_std

Trait R1CSVar

pub trait R1CSVar<F: Field> {
    type Value: Debug + Eq + Clone;

    // Required methods
    fn cs(&self) -> ConstraintSystemRef<F>;
    fn value(&self) -> Result<Self::Value, SynthesisError>;

    // Provided method
    fn is_constant(&self) -> bool { ... }
}

This trait describes some core functionality that is common to high-level variables, such as Booleans, FieldVars, GroupVars, etc.

Required Associated Types

type Value: Debug + Eq + Clone

The type of the “native” value that Self represents in the constraint system.

Required Methods

fn cs(&self) -> ConstraintSystemRef<F>

Returns the underlying ConstraintSystemRef.

If self is a constant value, then this must return ark_relations::r1cs::ConstraintSystemRef::None.

fn value(&self) -> Result<Self::Value, SynthesisError>

Returns the value that is assigned to self in the underlying ConstraintSystem.

Provided Methods

fn is_constant(&self) -> bool

Returns true if self is a circuit-generation-time constant.