Crate ark_r1cs_std

This crate implements common “gadgets” that make programming rank-1 constraint systems easier.

Re-exports

pub use self::bits::*;

Modules

alloc

This module describes a trait for allocating new variables in a constraint system.

bits

This module implements gadgets related to bit manipulation, such as Boolean and UInts.

eq

This module describes a trait for checking equality of variables.

fields

This module implements gadgets related to field arithmetic.

groups

This module implements gadgets related to group arithmetic, and specifically elliptic curve arithmetic.

macros

Some utility macros for making downstream impls easier.

pairing

This module implements gadgets related to computing pairings in bilinear groups.

poly

This module implements functions for manipulating polynomial variables over finite fields.

prelude

select

This module describes traits for conditionally selecting a variable from a list of variables.

Macros

impl_bounded_ops

Implements arithmetic traits (eg: Add, Sub, Mul) for the given type using the impl in $impl.

impl_ops

Implements arithmetic traits (eg: Add, Sub, Mul) for the given type using the impl in $impl.

Traits

Assignment

A utility trait to convert Self to Result<T, SynthesisErrorA.>

R1CSVar

This trait describes some core functionality that is common to high-level variables, such as Booleans, FieldVars, GroupVars, etc.

ToConstraintFieldGadget

Specifies how to convert a variable of type Self to variables of type FpVar<ConstraintF>