ark_r1cs_std

Module groups

This module implements gadgets related to group arithmetic, and specifically elliptic curve arithmetic.

Re-exports

pub use self::curves::short_weierstrass::bls12;

pub use self::curves::short_weierstrass::mnt4;

pub use self::curves::short_weierstrass::mnt6;

Modules

curves

This module contains implementations of arithmetic for various curve models.

Traits

CurveVar

A variable that represents a curve point for the curve C.

GroupOpsBounds

A hack used to work around the lack of implied bounds.